move: 도메인에 연관된 헤더를 common으로 이동하기
